requirements.txt
pytest~=8.3.5
allure-pytest~=2.14.0
selenium~=4.31.0
